io_uring: track status of fixed files with a list
From: Hao Xu <howeyxu@tencent.com>

Add a list to link all free fixed file entryies and some helpers to
update the list in time when a fixed file slot is used or freed.
This is helpful for functionality of finding a slot to allocate when
users ask to do so.

Signed-off-by: Hao Xu <howeyxu@tencent.com>