io_uring: allow allocated fixed files for openat/openat2/accept
From: Hao Xu <howeyxu@tencent.com>

If the applications passes in UINT_MAX as the file_slot, then that's a
hint to allocate a fixed file descriptor rather than have one be passed
in directly.

This can be useful for having io_uring manage the direct descriptor space.

Normal open/accept direct requests will complete with 0 for success,
and < 0 in case of error. If io_uring is asked to allocated the direct
descriptor, then the direct descriptor is returned in case of success.

Signed-off-by: Hao Xu <howeyxu@tencent.com>